Split user methods into new UserController
