+++ app/Http/Controllers/TableController.php
@@ -9,7 +9,6 @@ use App\Handlers\CreateTableHandler;
 use App\Handlers\UpdateTableHandler;
 use App\Http\Requests\TableStoreRequest;
 use App\Logic\TableLogic;
-use App\Logic\UserLogic;
 use App\Models\Category;
 use App\Models\Day;
 use App\Models\Game;
@@ -18,7 +17,6 @@ use App\Notifications\Discord\NotificationFactory;
 use App\Repositories\GameRepository;
 use Exception;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Log;
 use Illuminate\View\View;
@@ -114,44 +112,6 @@ class TableController extends Controller
         }
 
         return redirect()->route('days.show', $command->table->day);
-    }
-
-    public function subscribe(Table $table): RedirectResponse
-    {
-        if (app(UserLogic::class)->hasSubscribedToAnotherTableWithTheSameStartHour($table->day, $table)) {
-            return redirect()->route('days.show', $table->day)->with(['error' => 'Vous êtes déjà inscrit à une autre table à la même heure ce jour là']);
-        }
-
-        if (app(UserLogic::class)->isAlreadySubscribedToATable($table)) {
-            return redirect()
-                ->route('days.show', $table->day)
-                ->with([
-                    'error' => 'Vous êtes déjà inscrit à cette table',
-                ]);
-        }
-
-        $table->users()->attach(Auth::user());
-
-        $discordNotificationData = $this->discordNotificationData::make($table->game, $table, $table->day);
-
-        $discordNotification = ($this->notificationFactory)(entity: 'user', type: 'user-subscription',
-            discordNotificationData: $discordNotificationData);
-        $discordNotification->handle();
-
-        return redirect()->back();
-    }
-
-    public function unSubscribe(Table $table): RedirectResponse
-    {
-        $table->users()->detach(Auth::user());
-
-        $discordNotificationData = $this->discordNotificationData::make($table->game, $table, $table->day);
-
-        $discordNotification = ($this->notificationFactory)(entity: 'user', type: 'user-unsubscription',
-            discordNotificationData: $discordNotificationData);
-        $discordNotification->handle();
-
-        return redirect()->back();
     }
 
     public function destroy(Table $table): RedirectResponse
